+++ tests/test_njitted.py
@@ -1,3 +1,4 @@
+import pytest
 import numpy as np
 from numpy.testing import assert_allclose
 
@@ -6,7 +7,12 @@ from .test_utils import get_h
 from emg3d import solver, njitted, utils
 
 
-def test_amat_x():
+@pytest.mark.parametrize("njit", [True, False])
+def test_amat_x(njit):
+    if njit:
+        amat_x = njitted.amat_x
+    else:
+        amat_x = njitted.amat_x.py_func
 
     # 1. Compare to alternative amat_x
 
@@ -37,10 +43,9 @@ def test_amat_x():
 
     # amat_x
     rr1 = utils.Field(grid)
-    njitted.amat_x(
-            rr1.fx, rr1.fy, rr1.fz, efield.fx, efield.fy, efield.fz,
-            vmodel.eta_x, vmodel.eta_y, vmodel.eta_z, vmodel.zeta, grid.hx,
-            grid.hy, grid.hz)
+    amat_x(rr1.fx, rr1.fy, rr1.fz, efield.fx, efield.fy, efield.fz,
+           vmodel.eta_x, vmodel.eta_y, vmodel.eta_z, vmodel.zeta, grid.hx,
+           grid.hy, grid.hz)
 
     # amat_x - alternative
     rr2 = utils.Field(grid)
@@ -75,7 +80,19 @@ def test_amat_x():
     #     - Problem3D_e => return C.T*MfMui*C + 1j*omega(freq)*MeSigma
 
 
-def test_gauss_seidel():
+@pytest.mark.parametrize("njit", [True, False])
+def test_gauss_seidel(njit):
+    if njit:
+        gauss_seidel = njitted.gauss_seidel
+        gauss_seidel_x = njitted.gauss_seidel_x
+        gauss_seidel_y = njitted.gauss_seidel_y
+        gauss_seidel_z = njitted.gauss_seidel_z
+    else:
+        gauss_seidel = njitted.gauss_seidel.py_func
+        gauss_seidel_x = njitted.gauss_seidel_x.py_func
+        gauss_seidel_y = njitted.gauss_seidel_y.py_func
+        gauss_seidel_z = njitted.gauss_seidel_z.py_func
+
     # At the moment we only compare `gauss_seidel_x/y/z` to `gauss_seidel`.
     # Better tests should be implemented.
 
@@ -89,8 +106,8 @@ def test_gauss_seidel():
         # `gauss_seidel`/`_x/y/z` loop over z, then y, then x. Together with
         # `lr_dir`, we have to keep the dimension at 2 in order that they
         # agree.
-        nx = [4, 1, 1][lr_dir-1]
-        ny = [4, 1, 1][lr_dir-1]
+        nx = [1, 4, 4][lr_dir-1]
+        ny = [4, 1, 4][lr_dir-1]
         nz = [4, 4, 1][lr_dir-1]
 
         # Get this grid.
@@ -121,21 +138,27 @@ def test_gauss_seidel():
 
         # Get result from `gauss_seidel`.
         cfield = utils.Field(grid, efield.copy())
-        njitted.gauss_seidel_x(cfield.fx, cfield.fy, cfield.fz, *inp)
+        gauss_seidel(cfield.fx, cfield.fy, cfield.fz, *inp)
 
         # Get result from `gauss_seidel_x/y/z`.
         if lr_dir == 1:
-            njitted.gauss_seidel_x(efield.fx, efield.fy, efield.fz, *inp)
+            gauss_seidel_x(efield.fx, efield.fy, efield.fz, *inp)
         elif lr_dir == 2:
-            njitted.gauss_seidel_y(efield.fx, efield.fy, efield.fz, *inp)
+            gauss_seidel_y(efield.fx, efield.fy, efield.fz, *inp)
         elif lr_dir == 3:
-            njitted.gauss_seidel_z(efield.fx, efield.fy, efield.fz, *inp)
+            gauss_seidel_z(efield.fx, efield.fy, efield.fz, *inp)
 
         # Check the resulting field.
         assert_allclose(efield.field, cfield.field)
 
 
-def test_solve():
+@pytest.mark.parametrize("njit", [True, False])
+def test_solve(njit):
+    if njit:
+        solve = njitted.solve
+    else:
+        solve = njitted.solve.py_func
+
     # Create real and complex symmetric matrix A in the form as required by
     # njitted.solve, hence only main diagonal and five lower diagonals in a
     # vector.
@@ -181,7 +204,7 @@ def test_solve():
         # main diagonal and first five lower off-diagonals, arranged in a
         # vector one after the other. Convert matrix A into required vector A.
         res1 = b.copy()
-        njitted.solve(avec.copy(), res1)               # Solve A x = b
+        solve(avec.copy(), res1)               # Solve A x = b
         assert_allclose(x, res1)                       # Check
 
         # 3. Compare to alternative solver
@@ -198,24 +221,22 @@ def test_solve():
         assert_allclose(x, res2)                       # Check
 
 
-def test_restrict():
+@pytest.mark.parametrize("njit", [True, False])
+def test_restrict(njit):
+    if njit:
+        restrict = njitted.restrict
+        restrict_weights = njitted.restrict_weights
+    else:
+        restrict = njitted.restrict.py_func
+        restrict_weights = njitted.restrict_weights.py_func
+
     # Simple comparison using the most basic mesh.
-    h = np.array([1, 1, 1, 1])
+    h = np.array([1, 1, 1, 1, 1, 1])
 
-    # Fine and coarse grid.
-    fgrid = utils.TensorMesh([h, h, h], x0=np.array([-2, -2, -2]))
-    cgrid = utils.TensorMesh([np.diff(fgrid.vectorNx[::2]),
-                              np.diff(fgrid.vectorNy[::2]),
-                              np.diff(fgrid.vectorNz[::2])],
-                             fgrid.x0)
+    # Fine grid.
+    fgrid = utils.TensorMesh([h, h, h], x0=np.array([-3, -3, -3]))
 
-    # Regular grid, so all weights (wx, wy, wz) are the same...
-    w = njitted.restrict_weights(
-        fgrid.vectorNx, fgrid.vectorCCx, fgrid.hx, cgrid.vectorNx,
-        cgrid.vectorCCx, cgrid.hx)
-
-    # Create fields.
-    cfield = utils.Field(cgrid)
+    # Create fine field.
     ffield = utils.Field(fgrid)
 
     # Put in values.
@@ -223,11 +244,32 @@ def test_restrict():
     ffield.fy[:, :, :] = 2
     ffield.fz[:, :, :] = 4
 
-    # Ensure PEC and restrict them.
+    # Ensure PEC.
     ffield.ensure_pec
-    njitted.restrict(cfield.fx, cfield.fy, cfield.fz,
-                     ffield.fx, ffield.fy, ffield.fz,
-                     w, w, w, 0)
+
+    # Get weigths
+    wlr = np.zeros(fgrid.nNx, dtype=float)
+    w0 = np.ones(fgrid.nNx, dtype=float)
+    fw = (wlr, w0, wlr)
+
+    # # CASE 0 -- regular # #
+
+    # Coarse grid.
+    cgrid = utils.TensorMesh([np.diff(fgrid.vectorNx[::2]),
+                              np.diff(fgrid.vectorNy[::2]),
+                              np.diff(fgrid.vectorNz[::2])],
+                             fgrid.x0)
+
+    # Regular grid, so all weights (wx, wy, wz) are the same...
+    w = restrict_weights(fgrid.vectorNx, fgrid.vectorCCx, fgrid.hx,
+                         cgrid.vectorNx, cgrid.vectorCCx, cgrid.hx)
+
+    # Create coarse field.
+    cfield = utils.Field(cgrid)
+
+    # Restrict it.
+    restrict(cfield.fx, cfield.fy, cfield.fz, ffield.fx, ffield.fy, ffield.fz,
+             w, w, w, 0)
 
     # Check sum of fine and coarse fields.
     assert cfield.fx.sum() == ffield.fx.sum()
@@ -238,8 +280,141 @@ def test_restrict():
     np.allclose(cfield.fx[0, :, :]*2, cfield.fy[:, 0, :])
     np.allclose(cfield.fy[:, 0, :]*2, cfield.fz[:, :, 0])
 
+    # # CASE 1 -- y & z # #
 
-def test_restrict_weights():
+    # Coarse grid.
+    cgrid = utils.TensorMesh([fgrid.hx, np.diff(fgrid.vectorNy[::2]),
+                              np.diff(fgrid.vectorNz[::2])], fgrid.x0)
+
+    # Create coarse field.
+    cfield = utils.Field(cgrid)
+
+    # Restrict field.
+    restrict(cfield.fx, cfield.fy, cfield.fz, ffield.fx, ffield.fy, ffield.fz,
+             fw, w, w, 1)
+
+    # Check sum of fine and coarse fields.
+    assert cfield.fx.sum() == ffield.fx.sum()
+    assert cfield.fy.sum() == ffield.fy.sum()
+    assert cfield.fz.sum() == ffield.fz.sum()
+
+    # Assert fields are multiples from each other.
+    np.allclose(cfield.fy[:, 0, :]*2, cfield.fz[:, :, 0])
+
+    # # CASE 2 -- x & z # #
+
+    # Coarse grid.
+    cgrid = utils.TensorMesh([np.diff(fgrid.vectorNx[::2]), fgrid.hy,
+                              np.diff(fgrid.vectorNz[::2])], fgrid.x0)
+
+    # Create coarse field.
+    cfield = utils.Field(cgrid)
+
+    # Restrict field.
+    restrict(cfield.fx, cfield.fy, cfield.fz, ffield.fx, ffield.fy, ffield.fz,
+             w, fw, w, 2)
+
+    # Check sum of fine and coarse fields.
+    assert cfield.fx.sum() == ffield.fx.sum()
+    assert cfield.fy.sum() == ffield.fy.sum()
+    assert cfield.fz.sum() == ffield.fz.sum()
+
+    # Assert fields are multiples from each other.
+    np.allclose(cfield.fx[0, :, :].T*2, cfield.fz[:, :, 0])
+
+    # # CASE 3 -- x & y # #
+
+    # Coarse grid.
+    cgrid = utils.TensorMesh([np.diff(fgrid.vectorNx[::2]),
+                              np.diff(fgrid.vectorNy[::2]), fgrid.hz],
+                             fgrid.x0)
+
+    # Create coarse field.
+    cfield = utils.Field(cgrid)
+
+    # Restrict field.
+    restrict(cfield.fx, cfield.fy, cfield.fz, ffield.fx, ffield.fy, ffield.fz,
+             w, w, fw, 3)
+
+    # Check sum of fine and coarse fields.
+    assert cfield.fx.sum() == ffield.fx.sum()
+    assert cfield.fy.sum() == ffield.fy.sum()
+    assert cfield.fz.sum() == ffield.fz.sum()
+
+    # Assert fields are multiples from each other.
+    np.allclose(cfield.fx[0, :, :]*2, cfield.fy[:, 0, :])
+
+    # # CASE 4 -- x # #
+
+    # Coarse grid.
+    cgrid = utils.TensorMesh(
+            [np.diff(fgrid.vectorNx[::2]), fgrid.hy, fgrid.hz], fgrid.x0)
+
+    # Create coarse field.
+    cfield = utils.Field(cgrid)
+
+    # Restrict field.
+    restrict(cfield.fx, cfield.fy, cfield.fz, ffield.fx, ffield.fy, ffield.fz,
+             w, fw, fw, 4)
+
+    # Check sum of fine and coarse fields.
+    assert cfield.fx.sum() == ffield.fx.sum()
+    assert cfield.fy.sum() == ffield.fy.sum()
+    assert cfield.fz.sum() == ffield.fz.sum()
+
+    # Assert fields are multiples from each other.
+    np.allclose(cfield.fy[:, 0, :]*2, cfield.fz[:, :, 0])
+
+    # # CASE 5 -- y # #
+
+    # Coarse grid.
+    cgrid = utils.TensorMesh(
+            [fgrid.hx, np.diff(fgrid.vectorNy[::2]), fgrid.hz], fgrid.x0)
+
+    # Create coarse field.
+    cfield = utils.Field(cgrid)
+
+    # Restrict field.
+    restrict(cfield.fx, cfield.fy, cfield.fz, ffield.fx, ffield.fy, ffield.fz,
+             fw, w, fw, 5)
+
+    # Check sum of fine and coarse fields.
+    assert cfield.fx.sum() == ffield.fx.sum()
+    assert cfield.fy.sum() == ffield.fy.sum()
+    assert cfield.fz.sum() == ffield.fz.sum()
+
+    # Assert fields are multiples from each other.
+    np.allclose(cfield.fx[0, :, :].T*4, cfield.fz[:, :, 0])
+
+    # # CASE 6 -- z # #
+
+    # Coarse grid.
+    cgrid = utils.TensorMesh(
+            [fgrid.hx, fgrid.hy, np.diff(fgrid.vectorNz[::2])], fgrid.x0)
+
+    # Create coarse field.
+    cfield = utils.Field(cgrid)
+
+    # Restrict field.
+    restrict(cfield.fx, cfield.fy, cfield.fz, ffield.fx, ffield.fy, ffield.fz,
+             fw, fw, w, 6)
+
+    # Check sum of fine and coarse fields.
+    assert cfield.fx.sum() == ffield.fx.sum()
+    assert cfield.fy.sum() == ffield.fy.sum()
+    assert cfield.fz.sum() == ffield.fz.sum()
+
+    # Assert fields are multiples from each other.
+    np.allclose(cfield.fx[0, :, :]*4, cfield.fy[:, 0, :])
+
+
+@pytest.mark.parametrize("njit", [True, False])
+def test_restrict_weights(njit):
+    if njit:
+        restrict_weights = njitted.restrict_weights
+    else:
+        restrict_weights = njitted.restrict_weights.py_func
+
     # 1. Simple example following equation 9, [Muld06]_.
     edges = np.array([0., 500, 1200, 2000, 3000])
     width = (edges[1:]-edges[:-1])
@@ -254,7 +429,7 @@ def test_restrict_weights():
     wtr = np.array([350/600, 500/900, 400/500])
 
     # Result from implemented function
-    wl, w0, wr = njitted.restrict_weights(
+    wl, w0, wr = restrict_weights(
             edges, centr, width, c_edges, c_centr, c_width)
 
     assert_allclose(wtl, wl)
@@ -281,7 +456,7 @@ def test_restrict_weights():
             cgrid.vectorCCx, cgrid.hx)
 
     # Get the implemented numba-result
-    wxl, wx0, wxr = njitted.restrict_weights(
+    wxl, wx0, wxr = restrict_weights(
             grid.vectorNx, grid.vectorCCx, grid.hx, cgrid.vectorNx,
             cgrid.vectorCCx, cgrid.hx)
 
@@ -291,7 +466,13 @@ def test_restrict_weights():
     assert_allclose(wxr, wr)
 
 
-def test_blocks_to_amat():
+@pytest.mark.parametrize("njit", [True, False])
+def test_blocks_to_amat(njit):
+    if njit:
+        blocks_to_amat = njitted.blocks_to_amat
+    else:
+        blocks_to_amat = njitted.blocks_to_amat.py_func
+
     # Create some dummy data
 
     # Main matrix amat and RHS bvec
@@ -320,9 +501,9 @@ def test_blocks_to_amat():
     rhs3 = np.arange(11, 16)
 
     # Call blocks_to_amat three times to fill in all blocks
-    njitted.blocks_to_amat(amat, bvec, middle1, left1, rhs1, 0, 3)
-    njitted.blocks_to_amat(amat, bvec, middle2, left2, rhs2, 1, 3)
-    njitted.blocks_to_amat(amat, bvec, middle3, left3, rhs3, 2, 3)
+    blocks_to_amat(amat, bvec, middle1, left1, rhs1, 0, 3)
+    blocks_to_amat(amat, bvec, middle2, left2, rhs2, 1, 3)
+    blocks_to_amat(amat, bvec, middle3, left3, rhs3, 2, 3)
 
     # The result should look like this
     amat_res = np.arange(1., 91)
@@ -339,7 +520,13 @@ def test_blocks_to_amat():
     assert_allclose(bvec_res, bvec)
 
 
-def test_volume_average():
+@pytest.mark.parametrize("njit", [True, False])
+def test_volume_average(njit):
+    if njit:
+        volume_average = njitted.volume_average
+    else:
+        volume_average = njitted.volume_average.py_func
+
     # Comparison to alt_version.
     grid_in = utils.TensorMesh(
             [np.ones(30), np.ones(20)*5, np.ones(10)*10],
@@ -359,7 +546,7 @@ def test_volume_average():
 
     # New solution.
     new_values = np.zeros(grid_out.vnC, dtype=values.dtype)
-    njitted.volume_average(*points, values, *new_points, new_values, vol)
+    volume_average(*points, values, *new_points, new_values, vol)
 
     # Old solution.
     new_values_alt = np.zeros(grid_out.vnC, dtype=values.dtype)
@@ -367,3 +554,54 @@ def test_volume_average():
             *points, values, *new_points, new_values_alt)
 
     assert_allclose(new_values, new_values_alt)
+
+
+@pytest.mark.parametrize("njit", [True, False])
+def test_volume_avg_weights(njit):
+    if njit:
+        volume_avg_weights = njitted._volume_avg_weights
+    else:
+        volume_avg_weights = njitted._volume_avg_weights.py_func
+
+    grid_in = utils.TensorMesh(
+            [np.ones(11), np.ones(10)*2, np.ones(3)*10],
+            x0=np.array([0, 0, 0]))
+    grid_out = utils.TensorMesh(
+            [np.arange(4)+1, np.arange(5)+1, np.arange(6)+1],
+            x0=np.array([0.5, 3.33, 5]))
+
+    wx, ix_in, ix_out = volume_avg_weights(grid_in.vectorNx, grid_out.vectorNx)
+    assert_allclose(wx, [0, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 1, 1, 0.5, 0.5, 1, 1,
+                         1, 0.5, 0])
+    assert_allclose(ix_in, [0, 0, 1, 1, 2, 3, 3, 4, 5, 6, 6, 7, 8, 9, 10, 10])
+    assert_allclose(ix_out, [0, 0, 0, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3])
+
+    wy, iy_in, iy_out = volume_avg_weights(grid_in.vectorNy, grid_out.vectorNy)
+    assert_allclose(wy, [0, 0, 0.67, 0.33, 1.67, 0.33, 1.67, 1.33, 0.67, 2.,
+                         1.33, 0.67, 2, 2, 0.33, 0.])
+    assert_allclose(iy_in, [0, 1, 1, 2, 2, 3, 3, 4, 4, 5, 6, 6, 7, 8, 9, 9])
+    assert_allclose(iy_out, [0, 0, 0, 0, 1, 1, 2, 2, 3, 3, 3, 4, 4, 4, 4, 4])
+
+    wz, iz_in, iz_out = volume_avg_weights(grid_in.vectorNz, grid_out.vectorNz)
+    assert_allclose(wz, [0, 1, 2, 2, 1, 4, 5, 6, 0])
+    assert_allclose(iz_in, [0, 0, 0, 0, 1, 1, 1, 2, 2])
+    assert_allclose(iz_out, [0, 0, 1, 2, 2, 3, 4, 5, 5])
+
+    w, inp, out = volume_avg_weights(np.array([0., 5, 7, 10]),
+                                     np.array([-1., 1, 4, 6, 7, 11]))
+    assert_allclose(w, [1, 1, 3, 1, 1, 1, 3, 1])
+    assert_allclose(inp, [0, 0, 0, 0, 1, 1, 2, 2])
+    assert_allclose(out, [0, 0, 1, 2, 2, 3, 4, 4])
+
+
+@pytest.mark.parametrize("njit", [True, False])
+def test_l2norm(njit):
+    if njit:
+        l2norm = njitted.l2norm
+    else:
+        l2norm = njitted.l2norm.py_func
+
+    x = 1e-15*(np.arange(101.)-50)
+    compare = np.linalg.norm(x)
+    test = l2norm(x)
+    assert test == compare